Only remove if file is existing

+++ setup.sh
@@ -56,7 +56,10 @@ tar -xpf ${BASE_DIR}/tools/benchmark-kieker-instrumented/build/distributions/ben
 
 
 echo "Generating javassist load time instrumented benchmark"
-rm -r benchmark-kieker-javassist
+if [ -d benchmark-kieker-javassist ]
+then
+	rm -r benchmark-kieker-javassist
+fi
 cp benchmark benchmark-kieker-javassist -R
 java -cp ../kieker/build/libs/kieker-2.0.0-SNAPSHOT-javassist.jar kieker.monitoring.probe.javassist.BuildTimeAdaption benchmark-kieker-javassist/lib/benchmark.jar
 cd frameworks/Kieker-java-javassist-buildtime/
